re #124 separating new sys test to new fixture
