+++ system-tests/test_idle_pv_updates.py
@@ -3,7 +3,6 @@ from helpers.f142_logdata import LogData, Value, Double
 from helpers.flatbuffer_helpers import check_message_pv_name_and_value_type, check_double_value_and_equality
 from time import sleep
 from PVs import PVDOUBLE
-from pytest import raises
 
 
 def test_forwarder_sends_idle_pv_updates(docker_compose_idle_updates):
@@ -16,18 +15,3 @@ def test_forwarder_sends_idle_pv_updates(docker_compose_idle_updates):
         log_data_first = LogData.LogData.GetRootAsLogData(msg, 0)
         check_message_pv_name_and_value_type(log_data_first, Value.Value.Double, PVDOUBLE)
         check_double_value_and_equality(log_data_first, 0)
-
-
-def test_forwarder_does_not_send_pv_update_more_than_once_when_periodic_update_is_used(docker_compose_idle_updates_long_period):
-    consumer = create_consumer()
-    data_topic = "TEST_forwarderData_idle_updates"
-    consumer.subscribe([data_topic])
-    sleep(3)
-    msg = poll_for_valid_message(consumer).value()
-    log_data_first = LogData.LogData.GetRootAsLogData(msg, 0)
-    check_message_pv_name_and_value_type(log_data_first, Value.Value.Double, PVDOUBLE)
-    check_double_value_and_equality(log_data_first, 0)
-
-    with raises(AssertionError):
-        # AssertionError because there are no more messages to poll
-        msg = poll_for_valid_message(consumer).value()
